feat: API for updating product stock
Controller, service and repository logic. The SQL logic can probably be improved at a later point so that we don't make multiple queries inside the loop. For one thing we can probably use a subquery to fetch the current stock of the product instead of it being a separate query